item image v1 + item catalogue ui v1
- containerised item microservice (docker)
- created simple html bootstrap UI for item catalogue (will be exploring VUE soon)
