+++ docker/item/requirements.txt
@@ -1,0 +1,3 @@
+Flask==2.0.2
+Flask-SQLAlchemy==2.5.1
+mysql-connector-python==8.0.28
