+++ ui/index.html
@@ -1,0 +1,121 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+    <meta name="viewport" content="width=device-width">
+
+    <title>Item Catalogue </title>
+
+    <link rel="stylesheet" href="">
+    <!--[if lt IE 9]>
+    <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <!-- Bootstrap libraries -->
+    <meta name="viewport" 
+        content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+    <!-- Bootstrap CSS -->
+    <link href='https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css'
+    rel='stylesheet'
+    integrity='sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC'
+    crossorigin='anonymous'>
+
+    <!-- Latest compiled and minified JavaScript -->
+    <script
+    src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+
+</head>
+
+<body>
+
+    <div id="main-container" class="container">
+        <h1 class="display-4">Items on Sale!</h1>
+        <table class='table table-striped border-1'>
+            <thead class='table-dark'>
+                <tr>
+                    <th>ItemID</th>
+                    <th>ItemName</th>
+                    <th>SellerID</th>
+                    <th>ItemDesc</th>
+                    <th>Category</th>
+                    <th>Price</th>
+                    <th>ItemStatus</th>
+                </tr>
+            </thead>
+        <tbody id="itemTable"></tbody>
+        </table>
+        <a id="addItemBtn" class="btn btn-primary" href="add-item.html">Add an item</a>
+    </div>
+
+    <script>
+            // Helper function to display error message
+    function showError(message) {
+        // Hide the table and button in the event of error
+        $('#itemTable').hide();
+        $('#addItemBtn').hide();
+
+        // Display an error under the main container
+        $('#main-container')
+            .append("<label>"+message+"</label>");
+    }
+
+    // anonymous async function 
+    // - using await requires the function that calls it to be async
+    $(async() => {           
+        // Change serviceURL to your own
+        // var serviceURL = "http://127.0.0.1:5000/item";
+        var serviceURL = "http://192.168.1.5:5000/item";
+
+        try {
+            const response =
+                await fetch(
+                serviceURL, { method: 'GET' }
+            );
+            const result = await response.json();
+                if (response.status === 200) {
+                    
+                // success case
+                var items = result.data.item; //the array is in items within data of the returned result
+                // for loop to setup all table rows with obtained items data
+                var rows = "";
+                for (const item of items) {
+                    eachRow =
+                            "<td>" + item.ItemID + "</td>" +
+                            "<td>" + item.ItemName + "</td>" +
+                            "<td>" + item.Seller_ID + "</td>" +
+                            "<td>" + item.ItemDesc + "</td>" +
+                            "<td>" + item.Category + "</td>" +
+                            "<td>" + item.Price + "</td>" +
+                            "<td>" + item.ItemStatus + "</td>";
+                    rows += "<tr>" + eachRow + "</tr>";
+                }
+                    // add all the rows to the table
+                    $('#itemTable').append(rows);
+                } else if (response.status == 404) {
+                    // No items
+                    showError(result.message);
+                } else {
+                    // unexpected outcome, throw the error
+                    throw response.status;
+                }
+            } catch (error) {
+                // Errors when calling the service; such as network error, 
+                // service offline, etc
+                showError
+                ('There is a problem retrieving item data, please try again later.<br />' + error);
+            } // error
+    });
+    </script>
+
+
+    <!-- Bootstrap Javascript; at the end of the <body> -->
+<script 
+src='https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js'
+integrity='sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM'
+crossorigin='anonymous'>
+</script>
+</body>
+
+</html>
